requirements.txt
# 商机监测智能体 - 项目依赖

# ===== 核心框架 =====
browser-use>=0.11.2
pydantic>=2.11.5

# ===== 配置管理 =====
PyYAML>=6.0.1
python-dotenv>=1.0.0

# ===== 图像处理 =====
Pillow>=10.0.0

# ===== CLI和进度条 =====
rich>=13.0.0

# ===== HTTP API 服务 =====
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0

# ===== Python版本要求 =====
# Python >= 3.11
